OL-11409 Adding tacp_datacenter module to the CP ansible support.
It only adds the ability to create datacenters, as there is not API endpoint for deleting them; thus changes are not rolled back when a failure occurs, and a message to that effect is returned to the user upon failure so they are aware that the datacenter still exists, but not in the state that they desired.

It permits the assignment of migration zones, storage pools, and networks, as well as downloading a set of templates at creation time into the VDC Templates for that datacenter.

It also permits customized templates, or simply using all defaults by only providing the template's name. Can also specify to wait for templates to finish downloading before finishing the task's run. Design-wise there may be a wiser way to approach this, because if there's a mix of waits and don't-waits they do go in order, but I'm not sure how I could get them all to kick off at once, which would be the most efficient way I would think. I'm definitely open to ideas on that.